moved logic into composable to allow user to rewrite templates of components
