+++ src/components/QuestionTypes/MultiplePictureChoiceType.vue
@@ -5,17 +5,7 @@
     https://www.ditdot.hr/en
   */
 
-  import MultipleChoiceType from './MultipleChoiceType.vue'
-  import { QuestionType } from '../../models/QuestionModel'
+  import logic from '../../composables/QuestionTypes/MultiplePictureChoiceType'
 
-  export default {
-    extends: MultipleChoiceType,
-    name: QuestionType.MultiplePictureChoice,
-
-    data() {
-      return {
-        hasImages: true
-      }
-    }
-  }
+  export default logic
 </script>
