+++ src/components/QuestionTypes/PhoneType.vue
@@ -5,18 +5,7 @@
     https://www.ditdot.hr/en
   */
 
-  import TextType from './TextType.vue'
-  import LanguageModel from '../../models/LanguageModel'
-  import { QuestionType } from '../../models/QuestionModel'
-  
-  export default {
-    extends: TextType,
-    name: QuestionType.Phone,
-    data() {
-      return {
-        inputType: 'tel', 
-        canReceiveFocus: true
-      }
-    }
-  }
+  import logic from '../../composables/QuestionTypes/PhoneType'
+
+  export default logic
 </script>
